Update to use delta time as a Double and use RenderContext instead of raw JavaFX GC.

+++ src/main/kotlin/com/solsticesquared/hanabi/ParticleEmitter.kt
@@ -37,5 +37,5 @@ abstract class ParticleEmitter {
      * @param pool
      *        The pool of particle data to emit.
      */
-    abstract fun emit(dt: Float, pool: ParticlePool)
+    abstract fun emit(dt: Double, pool: ParticlePool)
 }
